Enhance project structure - Moved login dir to forms dir - Moved register dir to forms dir - Fixed imports for login/register to match new location - Removed unnecassary .keep files
