+++ src/js/components/forms/register/Register.ts
